Rename gdal_tile to tile_file so it reflects the class it contains.
